@@ -7,30 +7,32 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@emotion/react": "^11.8.1",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.4.2",
-    "@mui/lab": "^5.0.0-alpha.70",
-    "@mui/material": "^5.4.3",
-    "bson": "^4.6.1",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/lab": "5.0.0-alpha.100",
+    "@mui/material": "^5.8.7",
+    "axios": "^0.27.2",
     "date-fns": "^2.28.0",
-    "eslint-config-prettier": "^8.4.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-image-gallery": "^1.2.7",
-    "react-router-dom": "^6.2.2",
+    "immer": "^9.0.15",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-image-gallery": "^1.2.8",
+    "react-router-dom": "^6.3.0",
     "sanitize-html": "^2.7.0",
-    "swr": "^1.2.2"
+    "swr": "^1.3.0",
+    "zustand": "^4.1.1"
   },
   "devDependencies": {
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.6",
     "@types/sanitize-html": "^2.6.2",
-    "@typescript-eslint/eslint-plugin": "^5.12.1",
-    "@typescript-eslint/parser": "^5.12.1",
-    "@vitejs/plugin-react": "^1.2.0",
-    "eslint": "^8.9.0",
-    "typescript": "^4.5.5",
-    "vite": "^2.8.4"
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
+    "@vitejs/plugin-react": "^2.1.0",
+    "eslint": "^8.19.0",
+    "eslint-config-prettier": "^8.5.0",
+    "typescript": "^4.7.4",
+    "vite": "^3"
   }
 }
