@@ -1,14 +1,14 @@
 module.exports = {
   root: true,
   parser: '@typescript-eslint/parser',
-  plugins: [
-    '@typescript-eslint',
-  ],
+  plugins: ['@typescript-eslint'],
   rules: {
-    "@typescript-eslint/no-explicit-any": "off",
+    '@typescript-eslint/no-explicit-any': 'off',
+    '@typescript-eslint/no-non-null-assertion': 'off',
   },
   extends: [
     'eslint:recommended',
-    'plugin:@typescript-eslint/recommended', 'prettier',
+    'plugin:@typescript-eslint/recommended',
+    'prettier',
   ],
 }
