@@ -1,7 +1,6 @@
-import { ThemeProvider } from '@emotion/react'
-import { CssBaseline, useMediaQuery } from '@mui/material'
+import { CssBaseline, ThemeProvider, useMediaQuery } from '@mui/material'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import { FindPixivIllust } from './pages/FindPixivIllust'
+import { FindPixivIllust } from './pages/pixiv/Find'
 import { Layout } from './pages/Layout'
 import { NotFound } from './pages/NotFound'
 import { User as User } from './pages/pixiv/User'
