@@ -1,34 +1,26 @@
-import { ObjectId } from 'bson'
+// export interface LocalMedia<T> {
+//   _id: number
+//   url: string
+//   size: number
+//   mime: string
+//   local_path: string
+//   extension?: T
+// }
 
-export interface LocalMedia<T> {
-  _id: ObjectId
-  url: string
-  size: number
-  mime: string
-  local_path: string
-  extension?: T
-}
-
-export interface ImageMedia {
-  width: number
-  height: number
-  palette_hsv: Hsv[]
-}
-
-export interface Hsv {
-  h: number
-  s: number
-  v: number
-}
+// export interface ImageMedia {
+//   width: number
+//   height: number
+//   palette_hsv: Hsv[]
+// }
 
 export interface Item<E, H> {
-  _id: ObjectId
-  parent_id?: ObjectId
-  tag_ids?: ObjectId[]
+  id: number
+  parent_id?: number
+  tag_ids?: number[]
   source_id: string
   source_inaccessible: boolean
   last_modified?: Date
-  history: History<H>[]
+  history: History<H>
   extension?: E
 }
 
@@ -38,7 +30,6 @@ export interface History<H> {
 }
 
 export interface Tag {
-  _id: ObjectId
+  id: number
   alias: string[]
-  protected: boolean
 }
