@@ -1,9 +1,7 @@
 import { CircularProgress, Container } from '@mui/material'
-import { EJSON, ObjectId } from 'bson'
-import { useMemo } from 'react'
 import { useParams } from 'react-router-dom'
-import { PixivUser } from '../../model/pixiv'
-import { usePost } from '../../utils/network'
+import { PixivUser } from '../../../model/pixiv'
+import { usePost } from '../../../utils/network'
 
 export interface Params {
   id: string
@@ -12,18 +10,10 @@ export interface Params {
 export const User = () => {
   const { id } = useParams()
 
-  const oid = useMemo(() => {
-    try {
-      return new ObjectId(id)
-    } catch (error) {
-      return new ObjectId(0)
-    }
-  }, [id])
-
   const { data, error } = usePost<PixivUser[]>(
-    id !== undefined ? `/api/v1/pixiv/find/user` : null,
+    id !== undefined ? `/api/v2/pixiv/find/user` : null,
     {
-      ids: [oid],
+      ids: [id],
     }
   )
 
@@ -34,14 +24,14 @@ export const User = () => {
     if (error) {
       return error
     }
-    if (data && !data[0]) {
-      return 'User not found'
-    }
     if (!data) {
       return <CircularProgress />
     }
+    if (!data[0]) {
+      return 'User not found'
+    }
 
-    return EJSON.stringify(data)
+    return JSON.stringify(data)
   }
 
   return <Container>{render()}</Container>
